Fix dasht-server-http serving "400 Bad Request" on requests for "/"
This was reported in sunaku/dasht#51. It's not clear to me if this only affects
macOS users.

The problem was the `grep` query changed in this commit doesn't match queries
for "/":

    ~ $ cat urls
    /
    /?
    /?query
    /foo
    /bar
    ~ $ cat urls | grep '^/$\|^/?'
    /?
    /?query

The two patterns to match for are fairly simple, so they can instead be tested
for with simple string equality checks.